Fix fortify warning by refactoring to strcpy() + strcat()
gcc -D_FORTIFY_SOURCE=2 -O can detect some cases in which strncpy()
is called with its third parameter, "n", being calculated from the
source length. Parameter "n" should be based on the destination
buffer size: if "n" is based on the source length it can result in a
value greater than the size of the destination buffer and a buffer
overflow will occur. This will trigger a compiler warning.

This patch changes code in which is guaranteed that the source "s"
will always fit in the destination buffer "d". In this case, using
strncpy() is redundant anyway and a set of simpler strcpy() + strcat()
operations is used instead.